rotated rectangle
* rectangle with orientation-angle
* Point -> float precision
* MathHelper: round-function

+++ ImageProcessing/Point.cpp
@@ -1,13 +1,13 @@
 #include "Point.h"
 
 Point::Point() :
-  m_x(0),
-  m_y(0)
+  m_x(0.0),
+  m_y(0.0)
 {
 
 }
 
-Point::Point(unsigned int x, unsigned int y) :
+Point::Point(float x, float y) :
   m_x(x),
   m_y(y)
 {
